Fixed some input matching stuff
A long time ago when I made shooting and moving independent from each
other, I kinda switched the WASD keyboard input with the arrow keys.
Not a really big deal unless we try to do keybindings. Then it might
get confusing. So I went back and fixed all of it.