More heightmap and random world work
Heightmap problem has been fixed, it was a simple OutOfBounds error.
Added a spawner. Takes about 6-7 seconds on my computer to load the game
now with world gen. Loading screen will be necessary later. If you want
to test this, you will have to uncomment the transporter in PoorHouse
and import the level.