Swords, Horses, Trees and Decapitations! (STEPHEN AND DEREK READ BELOW)
For Derek

I made a tree entity and put it in Buildings. Size 17x32. Hardcolors:
(109401 and 593c01)
I switched Horses to their own sheet and made three variants:
Brown Horse: (4c301b and 202020)
Heavenly Horse: (f4f4f4 7cb2fe e3d002 and 202020)
Demonic Horse: (010101 810101 and 810101)
I added the sword swing animation tiles for the shortsword only. The
top row of Player_sword_sheet are the stabbing motions The bottom row
is the swing animation. This is a decapitating/powerful blow and should
work for any direction the player is in.

For Stephen
I guess I drew a cave wall tile and forgot about it, it’s on tile
sheet. (4,0). I coded it in to the tile class so you should be good to
go.