Spawning and Aesthetic Changes
Fixed spawning for random world and cave. There is no longer a chance of
spawning in the cave wall, and you always spawn on grass in random
world. Both spawns have been moved to the center of the levels. Also
changed how caves look, only the outermost wall is cave wall, the rest
is void. Dunno what you guys think, it can always be changed