String Formatter
Added a string formatter to the dialogue box that line breaks when the
String is too long. I left in a test String in the Jesus quest. If you
notice any text going over the edge, let me know what mob it is so I can
readjust the max length.