Font
It looks just like the one I made, I have the license. We have to
include it in the code, but there are no restrictions on using the font
in the game, as long as we don't sell the font exclusively, sorry
Stephen I know you wanted to do that, you're just going to have to make
your own.