Items and Colors
I believe it works, but it’s hard to tell through the flickering damn
you OSX! on the bright side, pause screen and inventory screen are both
semi-functional on macs. It flickers at a seizurely rate, but you can
click on everything. The guns still looked like red blocks, but I’ll
check it out in windows tomorrow