Update To Saves
Save file now includes Health data, Stamina data, and Score data as
well as Position data. I’ve come up with a good way to store
information. The text file where the data is stored can be hacked if
the user knows how to edit files in a jar file but I wouldn’t worry
about it. If that’s a worry, we can just encrypt the stuff with basic
encryption.