New Buildings Part 1
Added in a lot of new buildings, havn't touched interiors yet. I
included a building status text doc in res, which lists what has been
done. PLACED means that the building is in level1, INTERIOR means that
the building has the proper interior, and COLOR means that the building
needs a new color scheme.