A lot of Stuff
Still working on Unique Interiors, HippyVille is finished

Started work on World Map - Used actual in-game map and am coloring in
the cities with the tile colors. The MAP WILL NOT BE LOADED ALL AT
ONCE, don’t worry i know, this just gives me a basis to work off of for
every city, then we can copy the city area out and make it it’s own
level.

Some coloring fixes for the player and friendly sprites.

Also I discovered that when you color something black i.e. FF000000, it
renders invisible, just change it to FF111111 if you notice a
see-through object