Beginning of the Pathfinding algorithm
It will return an array, each array will contain an instruction, coded
from 0-7 that correlates to the eight directions that a character can
take. Once all instructions have been coded, it will be terminated with
an 8, so that other functions know when to stop looking for
instructions.

I am trying to take care of easy cases first, ones where the mob is only
moving in a straight line. I am hoping to check this for collisions and
if there are none, avoid the cost of running the whole algorithm.