GUI UPDATE
Added the Inventory Panels and Dialogue Screen meaning that our Beta
GUI is now finished!

Also added a GUI Female Player image for the GUI in preparation to add
in the female player sprites.

Also added are the temporary dialogue GUI panels and some templates so
that it’s clear how they are organized (these are temporary and subject
to change).

Also added a few more mobs, the skeleton, dog, and cat. Skeleton has a
sword attack implemented, will probably add a bow attack in as well so
we can have ranged and close quarters skeletons.

Finally I deleted the unused sprite assets, I still need to recalibrate
the rendering for stairs but other than that everything should be fine.