Mountain Update
I really wanted to sync an update when I was in the mountains last week
called mountain update but I wasn’t’ able to, so here’s this one.

I redrew the mountain sprites again, hopefully to more improvement, I
also redrew some of the wall tiles in the same manner.

I also remade portions and completed the first draft of the original
level, so we’re almost good for Beta.
 May have broken some quests because I deleted mobs.