A Whole Grab Bag of Updates
We got some sprite changes/updates, a new Evil Fox Character for the
Evil Fox Quest (is now an enemy so that his skeleton guard do not
attack him), added in Demons to the farm, as well as a Demon spawner
near the castle. I also created interiors for every building in the
Alpha, so everything is now enterable at the very least. I’m gonna sync
and test in Windows.