Let this be the Hour, When We Draw Swords Together
I rewrote the Sword Code to be less Stupid, which was only possible because I madae the Sword Sprite Sheets less Stupid. Now everything is absolutley uniform, every sword uses the same offset making it both easier to change things in the code later and also easier to change the sprites and identify issues quicker.
• Added new walking animations so that the player always keeps sword in same hand
• Reorganized Sprite Sheet, file size maintained
• Re-factored Player and Sword Class code, should make updating rendering ultimately easier
• Commented the hud out until performance issues can be dealt with

• Identified 5 Year Old Mac Bug caused by special mac keyboard upon holding down letter keys, SOLUTION: UPDATE JAVA VERSION TO MOST RECENT RELEASE